@@ -20,17 +20,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.9",
+    "@angular/animations": "^18.2.13",
     "@angular/cdk": "^17.3.9",
-    "@angular/common": "^17.3.9",
-    "@angular/compiler": "^17.3.9",
-    "@angular/core": "^17.3.9",
-    "@angular/forms": "^17.3.9",
+    "@angular/common": "^18.2.13",
+    "@angular/compiler": "^18.2.13",
+    "@angular/core": "^18.2.13",
+    "@angular/forms": "^18.2.13",
     "@angular/material": "^17.3.9",
     "@angular/material-date-fns-adapter": "^17.3.9",
-    "@angular/platform-browser": "^17.3.9",
-    "@angular/platform-browser-dynamic": "^17.3.9",
-    "@angular/router": "^17.3.9",
+    "@angular/platform-browser": "^18.2.13",
+    "@angular/platform-browser-dynamic": "^18.2.13",
+    "@angular/router": "^18.2.13",
     "@azure/msal-browser": "3.2.0",
     "@enigmatry/entry-components": "17.0.0",
     "@enigmatry/entry-form": "17.0.0",
@@ -40,13 +40,13 @@
     "date-fns": "^2.30.0",
     "rxjs": "^7.8.1",
     "tslib": "^2.3.0",
-    "zone.js": "~0.14.6"
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.7",
-    "@angular/cli": "^17.3.7",
-    "@angular/compiler-cli": "^17.3.9",
-    "@angular/localize": "^17.3.9",
+    "@angular/build": "^18.2.12",
+    "@angular/cli": "^18.2.12",
+    "@angular/compiler-cli": "^18.2.13",
+    "@angular/localize": "^18.2.13",
     "@enigmatry/eslint-config": "17.0.0",
     "@enigmatry/scss-foundation": "17.0.0",
     "@enigmatry/stylelint-config": "17.0.0",
